Working on moving to the haml-spec tests
